feature-develop: extracted components from home to shared and reuse them on projects
