+++ src/app/shared/components/about/about.component.scss
