+++ src/app/shared/components/projects/components/card/card.component.html
