+++ src/app/shared/components/projects/projects.component.scss
